bittu-bot/requirements.txt
discord.py==1.7.3
opencv-python-headless
asyncpraw==7.5.0
